Refactor assertDestroyablesDestroyed to match real implementation.
In order to avoid permanently leaking all destroyables in debug builds
the final implementation in `@glimmer/runtime` implemented
`assertDestroyablesDestroyed` to require a `enableDestroyableTracking`
method to be called first (this swaps the internal tracking from a
`WeakMap` to a `Map`). `enableDestroyableTracking` is expected to be
called before each test that will use `assertDestroyablesDestroyed` at
the end.