Ensure functionality with ember-cli-babel@7.22.0.
In ember-cli-babel@7.22.0 the imports from `@ember/destroyable` are
rewritten to be global accessors from `Ember.*`. And specifically,
`import { destroy } from '@ember/destroyable';` is rewritten to
`Ember.destroy`. This means that we need to properly handle the normal
`Ember.destroy` behavior (which changes in Ember 3.16.4+ and then again
in 3.20.0-beta.4+) along with our customized behavior (to run
destroyables).

This ensures that `Ember.destroy` is _our_ destroy method and runs run
destructors as needed.