Avoid errors when `super.init` is called multiple times.
When running on Ember < 3.20 calling the super classes `init` more than
once would issue an assertion:

```js
`'${destructor}' is already registered with '${destroyable}'.`
```

Normally, any case where you are actually invoking `super.init` multiple
times should just be fixed in your source (it is fundamentally "bad" and
is almost certainly a logic error in application code). Unfortunately,
there are cases today where you _accidentally_ call the super classes
`init` method. For example, when you call `this._super()` in a computed
property if that property did not exist on the super class the value of
the `this._super` property is whatever _happened_ to be the most recent
super wrapped function that was invoked. 😱

This ensures that calling `init` multiple times does **not** error (and
does not end up causing `willDestroy` to be called more than once) by
only registering the destructor if needed.