fix: find `ember-source` version of project (#136)
Currently, the version-checking library is passed the addon itself to
find the `ember-source` version of. This is not recommended in the
package's README:

https://github.com/ember-cli/ember-cli-version-checker#should-i-use-project-or-parent

This has caused issues with locating the correct version when used
within a mono-repo that contains packages using different Ember versions.

Passing `this.project` instead ensures that the polyfill is installed based on the host app's Ember version, as it should.